+++ R/nice_frequencies.R
@@ -1,8 +1,10 @@
 #' Table with frequency statistics
 #'
-#' @param data A data frame
+#' @param data Vector.
+#' @param grouping A grouping variable as a vector.
 #' @param label Set label for the variable name.
 #' @param show_missing If TRUE, adds a row for the number of missing values.
+#' @param show_percent If TRUE, adds a column for percentages.
 #' @param auto_labels If TRUE, variable names are taken from a label attribute.
 #' @param title Table title.
 #' @param footnote Table footnote.
@@ -11,10 +13,18 @@
 #' @return An html table with frequencies
 #' @examples
 #' nice_frequencies(mtcars[[11]])
+#' nice_frequencies(
+#'   mtcars$cyl, mtcars$am, 
+#'   label = "Cylinders", 
+#'   label_grouping = "Automatic"
+#' )
 #' @export
 nice_frequencies <- function(data,
+                             grouping = NULL,
                              label = NULL,
+                             label_grouping = NULL,
                              show_missing = TRUE,
+                             show_percent = TRUE,
                              auto_labels = TRUE,
                              title = NULL,
                              footnote = NULL,
@@ -22,47 +32,73 @@ nice_frequencies <- function(data,
                              ...) {
   
   on.exit(print_messages())
-  if (auto_labels && is.null(label)) label <- attr(data, "label") 
-  if (is.null(label)) label <- deparse(substitute(data))
-  if (is.null(title)) {
+  
+  useNA <- if (show_missing) "always" else "no"
+  
+  if (auto_labels && is.null(label)) 
+    label <- attr(data, "label")
+  if (auto_labels && is.null(label_grouping)) 
+    label_grouping <- attr(grouping, "label")
+  if (is.null(label)) 
+    label <- deparse(substitute(data))
+  if (is.null(label_grouping)) 
+    label_grouping <- deparse(substitute(grouping))
+  if (is.null(title) && is.null(grouping)) {
     title <- paste0("Frequency statistics of '", label, "'")
   }
   
-  #if (is.null(footnote)) 
-  #  footnote <- paste0("MAD is the median average deviation with a ",
-  #                     "consistency adjustment")
+  if (is.null(title) && !is.null(grouping)) {
+    title <- paste0(
+      "Frequency statistics of '", label, "' by '", label_grouping, "'"
+    )
+  }
   
   
-  #.filter <- sapply(data, is.numeric)
+  if (is.null(grouping)) {
+    tab <- table(data, useNA = useNA)
+    class(tab) <- NULL
+    out <- as.data.frame(tab)
+    names(out)[1] <- "Frequency"
+    if (show_percent)
+      out$Percent = round(out[[1]]/sum(out[[1]], na.rm = TRUE)*100)
+    spanner <- NULL
+    rn <- rownames(out)
+  } else {
+    tab <- table(data, grouping, useNA = useNA)
+    class(tab) <- NULL
+    out <- as.data.frame(tab)
+    out[ncol(out)] <- NULL
+    rn <- row.names(tab)
+    group_levels <- ncol(out)
   
-  #if (any(!.filter)) {
-  #  add_message(
-  #    "Some variables are not numeric and dropped from the analysis: ",
-  #    paste0(names(.filter)[!.filter], collapse = ", ")
-  #  )
-  #}
+    if (show_percent) {
+      perc <- lapply(out, \(.) round(./sum(., na.rm = TRUE)*100))
+      perc <- as.data.frame(perc)
+      names(perc) <- paste0(" ", names(out), " ")
+      
+      out <- cbind(out, perc)
+      
+    }
+    spanner <- list(
+      Frequency = 2:(group_levels + 1), 
+      Percent = (group_levels + 2):(2 * group_levels + 1)
+    )
+    names(spanner)[1] <- label_grouping#paste0("Frequency '", label_grouping, "'")
+  }
   
-  #data <- data[, .filter]
-  
-  #if (auto_labels) data <- rename_from_labels(data) 
-  
-  tab <- if (show_missing) table(data, useNA = "always") else table(data)
-  class(tab) <- NULL
-  out <- as.data.frame(tab)
-  names(out)[1] <- "Frequency"
-  
-  rn <- rownames(out)
   if (length(which(is.na(rn))) > 0) rn[which(is.na(rn))] <- "Missing"
+  if (length(which(rn == "NA.")) > 0) rn[which(rn == "NA.")] <- "Missing"
+
   
   rownames(out) <- NULL
   out <- cbind(Value = rn, out)
-  
-  out$Percent = round(out[[2]]/sum(out[[2]], na.rm = TRUE)*100)
+  names(out)[1] <- label
   
   out <- set_wmisc_attributes(out, 
     title = title,
     footnote = footnote,
-    file = file
+    file = file,
+    spanner = spanner
   )
   
   nice_table(out, ...)